Fix data-mangling during replace()
If replace() is called on text that spans lines, and the replacement is
longer than the match, then in some situations some of the following
text was deleted. This swaps out the naive String.prototype.replace()
call with one that uses String.prototype.slice() instead, and assembles
the output string manually.